refactor: split requests controller into lifecycle + pipeline (#2253)
