+++ tests/unit/api/controllers/test_requests_intake_pipeline.py
@@ -10,7 +10,7 @@ the integration suite.
 import pytest
 
 from synthorg.api.approval_store import ApprovalStore
-from synthorg.api.controllers.requests import process_intake_pipeline
+from synthorg.api.controllers.requests.pipeline import process_intake_pipeline
 from synthorg.api.state import AppState
 from synthorg.client.models import ClientRequest, RequestStatus, TaskRequirement
 from synthorg.client.simulation_state import ClientSimulationState
